Init arraylist utils for converting arraylist string to array string
